First pass on loading state for "generate new codes"
