+++ lib/plausible_web/components/generic.ex
@@ -2,7 +2,7 @@ defmodule PlausibleWeb.Components.Generic do
   @moduledoc """
   Generic reusable components
   """
-  use Phoenix.Component
+  use Phoenix.Component, global_prefixes: ~w(x-)
 
   @notice_themes %{
     yellow: %{
@@ -266,6 +266,28 @@ defmodule PlausibleWeb.Components.Generic do
 
     ~H"""
     <%= Phoenix.HTML.raw(@code) %>
+    """
+  end
+
+  attr :class, :any, default: ""
+
+  def spinner(assigns) do
+    ~H"""
+    <svg
+      class={["animate-spin h-4 w-4 text-indigo-500", @class]}
+      xmlns="http://www.w3.org/2000/svg"
+      fill="none"
+      viewBox="0 0 24 24"
+    >
+      <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4">
+      </circle>
+      <path
+        className="opacity-75"
+        fill="currentColor"
+        d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
+      >
+      </path>
+    </svg>
     """
   end
