+++ lib/plausible_web/templates/auth/user_settings.html.heex
@@ -422,11 +422,18 @@
     form_data={@conn.params}
     form_target={Routes.auth_path(@conn, :disable_2fa)}
     title="Disable Two-Factor Authentication?"
-    submit_label="Disable 2FA"
   >
     <:icon>
       <Heroicons.shield_exclamation class="h-6 w-6" />
     </:icon>
+    <:buttons>
+      <button
+        type="submit"
+        class="w-full inline-flex justify-center rounded-md border border-transparent shadow-sm px-4 py-2 bg-indigo-600 text-base font-medium text-white hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-700 sm:ml-3 sm:w-auto sm:text-sm"
+      >
+        Disable 2FA
+      </button>
+    </:buttons>
 
     <div class="text-sm mt-2">
       Once disabled, verification codes from the authenticator application and current recovery codes will become invalid. 2FA will have to be setup from the start.
@@ -452,13 +459,34 @@
     state_param="regenerate2FAOpen"
     form_data={@conn.params}
     form_target={Routes.auth_path(@conn, :generate_2fa_recovery_codes)}
-    onsubmit="var button = document.querySelector('#regenerate-2fa-modal button[type=submit]'); button.disabled = true; button.classList.add('animate-pulse')"
+    onsubmit="var group = document.querySelector('#buttons'); group.classList.add('submitting');"
     title="Generate New Recovery Codes?"
-    submit_label="Generate New Codes"
   >
     <:icon>
       <Heroicons.key class="h-6 w-6" />
     </:icon>
+
+    <:buttons>
+      <div x-data="{submitted: false}" class="group" id="buttons">
+        <.button
+          type="submit"
+          x-on:click="submitted = true"
+          class="w-52 group-[.submitting]:bg-gray-400 dark:group-[.submitting]:bg-gray-800"
+        >
+          <span x-show="!submitted">
+            Generate New Codes
+          </span>
+
+          <span
+            x-show="submitted"
+            style="display: none;"
+            class="inline-flex items-center text-white dark:text-gray-600"
+          >
+            <.spinner class="h-5 w-5 mr-2 text-white dark:text-gray-600" /> Generating Codes
+          </span>
+        </.button>
+      </div>
+    </:buttons>
 
     <div class="text-sm mt-2">
       If you generate new codes, the old ones will become invalid.
