+++ lib/plausible_web/templates/auth/verify_2fa_recovery_code.html.heex
@@ -17,7 +17,7 @@
             value: "",
             autocomplete: "off",
             class:
-              "font-medium shadow-sm focus:ring-indigo-500 focus:border-indigo-500 block w-full px-8 border-gray-300 dark:border-gray-500 dark:text-gray-200 dark:bg-gray-900 rounded-md",
+              "font-medium shadow-sm focus:ring-indigo-500 focus:border-indigo-500 block w-full px-2 border-gray-300 dark:border-gray-500 dark:text-gray-200 dark:bg-gray-900 rounded-md",
             maxlength: "10",
             placeholder: "Enter recovery code",
             required: "required"
